feat: Implement pagination over large paste collection
This change adds a "Next Page" button on the frontend if there are more
pastes after the current page. Users should use the browser back button
if they want to visit previous pages since we don't support a "Previous
Page" button.

The "next" url field is passed in both Jinja2 and the JSON response
body so clients can paginate pastes by just following the url.

The default page size (limit) is 100. The query string should pass a
millisecond epoch time in "cursor" to start the query from a particular
paste.